Move testing dir into app/utils
