+++ src/app/service/canvas.service.spec.ts
@@ -1,8 +1,8 @@
 import { CanvasService } from './canvas.service';
 import { TestBed } from '@angular/core/testing';
 import { GoogleRealtimeService } from './google-realtime.service';
-import { FakeGapi } from '../../testing/fake/fake-gapi';
-import { FakeGoogleRealtimeService } from '../../testing/fake/fake-google-realtime.service';
+import { FakeGapi } from '../utils/testing/fake-gapi';
+import { FakeGoogleRealtimeService } from '../utils/testing/fake-google-realtime.service';
 
 describe('Canvas service', () => {
   let cs: CanvasService;
