+++ src/app/service/google-realtime.service.spec.ts
@@ -1,9 +1,9 @@
 import { TestBed, async } from '@angular/core/testing';
 import { GoogleRealtimeService } from './google-realtime.service';
-import { FakeGapi } from '../../testing/fake/fake-gapi';
+import { FakeGapi } from '../utils/testing/fake-gapi';
 import { Http } from '@angular/http';
 import { GoogleDriveService } from './google-drive.service';
-import { FakeHttp } from '../../testing/fake/fake-http';
+import { FakeHttp } from '../utils/testing/fake-http';
 import { RouterTestingModule } from '@angular/router/testing';
 
 describe('Google Realtime Service', () => {
